I made some modifications to matrix-quality.
1) Method CalcSequenceDistrib

The most important modification is that I suppressed the use of
	matrix-scan -return distrib,
followed by
	 classfreq
(with two separate result files),

and replaced this by a single command, combining
	matrix-scan -return sites [...] | cut -f 8 | classfreq -v [...]

The reason is that the option -return distrib returns the whole distribution of scores, and does not allow to impose thresolds. Importantly, for some distributions (e.g. annotated sites) we want to add some thresholds like
	-uth rank 1
to obtain only one score per site.

2) Option -graph_topion

This allows to add some options to XYgraph for generating the distribution curves (NOT the ROC curves anymore)

3) Option -roc_option

Allows to add some options to the ROC curves
Typically, you want different options on the ROC curves than on the distributions (for example, teh X and Y limits are typically different).

4) I suppressed the concepts of "negative" and "positive" sets.
Instead,he user can specify as many sequence sets as necessary, with the option
	-seq seq_type seq_file

thus, if you want to call your negative set "negative", you simply type
	-seq negative neg_sequences.fasta
 but you can also decide to call it "intergenic" or "exonic" depending on what you consider a a negative control.

5) I added an option -roc_ref, to specify the reference set for the ROC curve. This is th distribution that will be used for the X axis of the ROC curve, and all other distributions are plotted on the Y axis. This is complementary with the removal of the concept of "negative" set. If you want to compare all distributions to the "exonic" distribution, you can type
	-seq exonic my_exons.fasta -roc_ref exonic

6) I added a distinction between two ranks, which can be very useful when you are scanning a sequence with multiple matrices :
	- rank per sequence (all matrices together)
	- rank per matrix and sequence (one rank for each matrix separately)

You can set threshold on any of those or on both.
For instance, you might desire to know the best score for each matrix on each sequence on a set. Then
	-uth rank_pm 1

Or you might search for the top 5 scores in each sequence, irrespective of the matrix which raised this score (e.g. for searching heterotypic modules of binding sites).
	-uth rank 5

Actually, this was necessary for the score distributions of multiple matrix permutations. Indeed, for multiple permutations, I generate a single file with all the permuted matrices, and then scan the sequences with these matrices. If one imposes the option -uth rank 1, one has a problem: each sequence returns only one site, the best among the 10 or 20 permuted matrices. With the option -uth rank_pm 1, one has exactly one hit per matrix and per sequence.

Note that when I use the option -uth rank_pm 1 on annotated sites which are larger than the matrix, the permuted distribution is well above the diagonal on the ROC) because we obtain an extreme value distribution: for each permuted matrix, we return the best position among the 10 to 20 possible positions in the extended sites. One should thus be cautious with this option.




Jacques van Helden <Jacques.van.Helden@ulb.ac.be>